+++ src/main/java/com/bashpile/BashpileMain.java
@@ -1,68 +1,83 @@
 package com.bashpile;
 
-import org.antlr.v4.runtime.CharStream;
-import org.antlr.v4.runtime.CharStreams;
-import org.antlr.v4.runtime.CommonTokenStream;
-import org.antlr.v4.runtime.tree.ParseTree;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import picocli.CommandLine;
 
 import java.io.FileInputStream;
+import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
+import java.util.concurrent.Callable;
 
-import static com.bashpile.ArrayUtils.arrayOf;
+import static com.bashpile.AntlrUtils.parse;
 
 /** Entry point into the program */
-public class BashpileMain {
+@CommandLine.Command(
+        name = "bashpile",
+        description = "Converts Bashpile lines to Bash"
+)
+public class BashpileMain implements Callable<Integer> {
 
     private static final Logger log = LogManager.getLogger(BashpileMain.class);
 
-    public static void main(String[] args) throws IOException {
-        processArgs(args);
+    @CommandLine.Option(names = {"-i", "--inputFile"})
+    private String inputFile;
+
+    private CommandLine commandLine;
+
+    public BashpileMain() {}
+
+    public BashpileMain(String inputFile) {
+        this.inputFile = inputFile;
     }
 
-    public static String[] processArg(String filename) throws IOException {
-        return processArgs(arrayOf(filename));
+    public static void main(String[] args) {
+        BashpileMain bashpile = new BashpileMain();
+        CommandLine argProcessor = new CommandLine(bashpile);
+        bashpile.setCommandLine(argProcessor);
+        System.exit(argProcessor.execute(args));
     }
 
-    public static String[] processArgs(String[] args) throws IOException {
-        // stream is either stdin or the first argument
-        InputStream is = System.in;
-        boolean argsExist = args.length > 0;
-        if (argsExist) {
-            String inputFile = args[0];
-            is = new FileInputStream(inputFile);
+    @Override
+    public Integer call() {
+        commandLine.usage(System.out);
+        return 1;
+    }
+
+    @CommandLine.Command(name = "execute", description = "Converts Bashpile lines to bash and executes them")
+    public int executeCommand() {
+        System.out.println(execute());
+        return 0;
+    }
+
+    public String execute() {
+        log.debug("In {}", System.getProperty("user.dir"));
+        String bashScript = "<stream unparsed>";
+        try {
+            bashScript = parse(getInputStream());
+            return CommandLineExecutor.run(bashScript).getLeft();
+        } catch (Throwable e) {
+            throw new BashpileUncheckedException(e, "Couldn't run `%s`.".formatted(bashScript));
         }
-
-        return parse(is);
     }
 
-    /** antlr calls */
-    private static String[] parse(InputStream is) throws IOException {
-        log.trace("Starting parse");
-        // lexer
-        CharStream input = CharStreams.fromStream(is);
-        BashpileLexer lexer = new BashpileLexer(input);
-        CommonTokenStream tokens = new CommonTokenStream(lexer);
-
-        // parser
-        BashpileParser parser = new BashpileParser(tokens);
-        ParseTree tree = parser.prog();
-
-        return applyBashpileLogic(tree);
+    @CommandLine.Command(name = "transpile", description = "Converts Bashpile lines to bash")
+    public int transpileCommand() throws IOException {
+        System.out.println(parse(getInputStream()));
+        return 0;
     }
 
-    private static String[] applyBashpileLogic(ParseTree tree) {
-        // visitor
-        try (BashpileVisitor bashpileLogic = new BashpileVisitor()) {
-            String bashScript = bashpileLogic.visit(tree);
-            String output = Shell.run(bashScript);
-            return output.split("\r?\n");
-        } catch (IOException e) {
-            log.error(e);
-            e.printStackTrace();
-            return ArrayUtils.arrayOf(e.getMessage());
+    private InputStream getInputStream() throws FileNotFoundException {
+        if (inputFile != null) {
+            return new FileInputStream(inputFile);
+        } else {
+            System.out.println("Enter your bashpile program, ending with a newline and EOF (ctrl-D).");
+            return System.in;
         }
+    }
+
+    public void setCommandLine(CommandLine commandLine) {
+        this.commandLine = commandLine;
     }
 }