+++ src/main/java/com/bashpile/BashpileUncheckedException.java
@@ -4,7 +4,11 @@ public class BashpileUncheckedException extends RuntimeException {
     public BashpileUncheckedException(String message) {
         super(message);
     }
-    public BashpileUncheckedException(Exception e) {
+    public BashpileUncheckedException(Throwable e) {
         super(e);
+    }
+
+    public BashpileUncheckedException(Throwable e, String message) {
+        super(message, e);
     }
 }