+++ src/main/java/com/bashpile/FailableStreamConsumer.java
@@ -7,12 +7,16 @@ import java.io.BufferedReader;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 
-/** Adapted from <a href="https://www.baeldung.com/run-shell-command-in-java">a tutorial</a>. */
-public class StreamGobbler implements Runnable {
+/**
+ * Applies consumer to each inputStream line.  Exceptions are hanled both in the stream and the consumer.
+ * <br>
+ * Adapted from <a href="https://www.baeldung.com/run-shell-command-in-java">a tutorial</a>.
+ */
+public class FailableStreamConsumer implements Runnable {
     private final InputStream inputStream;
     private final FailableConsumer<String, Exception> consumer;
 
-    public StreamGobbler(InputStream inputStream, FailableConsumer<String, Exception> consumer) {
+    public FailableStreamConsumer(InputStream inputStream, FailableConsumer<String, Exception> consumer) {
         this.inputStream = inputStream;
         this.consumer = consumer;
     }