+++ docs/changelog.md
@@ -1,1 +1,2 @@
-0.1.0 - Initial checkins.  Simple grammar, JUnit5 tests, Log4j2 logging
+* 0.1.0 - Initial checkins.  Simple grammar, JUnit5 tests, Log4j2 logging, Bash output, *nix command line class
+* 0.2.0 - Factored out translation to script language to support multiple shells